requirements.txt
streamlit>=1.30.0
pydicom>=2.4.0
Pillow>=10.0.0
numpy>=1.24.0
requests>=2.31.0
pywebview>=5.0.0
